chore: fix all clippy lints, remove 18 CI suppressions
Fixed violations: approx_constant (test floats near PI), useless_vec
(heap tests), identity_op (WAL recovery test), unused-variables
(plan_cache), unnecessary_cast (bench), cloned_ref_to_slice_refs (bench).

13 of the 18 suppressed lints had zero violations and were already stale.
CI now runs `cargo clippy -- -D warnings` with no suppression flags.

+++ crates/bench/benches/storage.rs
@@ -117,7 +117,7 @@ fn bench_btree_lookup(c: &mut Criterion) {
 
     // Pre-generate 100 keys so the timed loop isn't measuring Int() construction.
     let keys: Vec<Value> = (0..100)
-        .map(|i| Value::Int((i * 491 % N_ROWS as i64) as i64))
+        .map(|i| Value::Int(i * 491 % N_ROWS as i64))
         .collect();
 
     let mut idx: usize = 0;
